pyproject.toml
[tool.poetry]
name = "chat-bot-api"
version = "0.0.1"
description = "ITMO chat bot subsystem"
authors = ["No Name <no.name@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pydantic = {extras = ["dotenv"], version = "^1.10.13"}
dependency-injector = "^4.40.0"
yoyo-migrations = "^8.1.0"
aiocache = {version = "^0.12.1", extras = ["redis"]}
aiopg = "^1.3.3"
starlette-prometheus = "^0.9.0"
httpx = "^0.26.0"
opentelemetry-instrumentation-fastapi = "^0.43b0"
opentelemetry-instrumentation-logging = "^0.43b0"
opentelemetry-exporter-otlp = "^1.22.0"
opentelemetry-api = "^1.22.0"
opentelemetry-sdk = "^1.22.0"
jsf = "0.7.1"
uvicorn = "^0.27.1"
fastapi = "0.109.0"
aio-pika = "^9.4.0"
ujson = "^5.9.0"
redis = {extras = ["hiredis"], version = "^5.0.3"}
argon2-cffi = "^23.1.0"
pre-commit = "3.2.0"
selenium = "4.26.1"
aioboto3 = "13.2.0"
openai = "^1.99.6"


[tool.poetry.group.formatters.dependencies]
autoflake = "^1.4"
isort = "^5.10.1"
black = "^22.1.0"
docformatter = "^1.4"
add-trailing-comma = "^2.2.1"

[tool.poetry.group.sasts.dependencies]
itsdangerous = "^2.1.2"
safety = "^2.3.3"
bandit = "^1.7.1"

[tool.poetry.group.linters.dependencies]
flake8 = "^4.0.1"
mypy = "^0.931"
pylint = "^2.17.6"
ruff = "^0.1.8"

[tool.poetry.group.tests.dependencies]
pytest-asyncio = "^0.20.2"
coverage = "^6.3.2"
pytest-cov = "^4.0.0"
pytest-repeat = "^0.9.1"

[tool.poetry.group.tools.dependencies]
pytest-watcher = "^0.3.4"
pytest-sugar = "^0.9.7"

[tool.ruff]
line-length = 88

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401", "F403"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
